+++ src/lib.rs
@@ -1,36 +1,28 @@
-use serde::{Deserialize, Serialize};
-
 pub mod cli;
 pub mod comms;
+pub mod network;
 pub mod storage;
 
 pub mod tracing {
-    use tracing_subscriber::{EnvFilter, layer::SubscriberExt as _};
+    use tracing_subscriber::{
+        EnvFilter,
+        fmt::{self, format::FmtSpan},
+        layer::SubscriberExt as _,
+    };
 
     /// Initializes the tracing subscriber with the given environment filter.
     pub fn init(env_filter: &str) {
+        // Formats logs with for stdout
+        let fmt_layer = fmt::layer()
+            .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
+            .pretty();
+
         let env_filter = EnvFilter::from(env_filter);
         let subscriber = tracing_subscriber::registry()
             .with(env_filter)
-            .with(tracing_subscriber::fmt::layer());
+            .with(fmt_layer);
+
         tracing::subscriber::set_global_default(subscriber)
             .expect("Failed to set global default subscriber");
     }
 }
-
-pub mod behavior {
-    use libp2p::swarm::NetworkBehaviour;
-    use libp2p::{gossipsub, kad};
-
-    #[derive(NetworkBehaviour)]
-    pub struct PeerBehavior {
-        // request_response: request_response::cbor::Behaviour<PeerRequest, PeerResponse>,
-        pub kademlia: kad::Behaviour<kad::store::MemoryStore>,
-        pub gossipsub: gossipsub::Behaviour,
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct PeerRequest {}
-#[derive(Debug, Serialize, Deserialize)]
-pub struct PeerResponse {}
