+++ src/network/mod.rs
@@ -1,0 +1,2 @@
+pub mod behaviour;
+pub mod event;
